+++ OpenAI/Packages/com.openai.unity/Runtime/Threads/IncompleteDetails.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+using UnityEngine.Scripting;
+
+namespace OpenAI.Threads
+{
+    [Preserve]
+    public sealed class IncompleteDetails
+    {
+        [Preserve]
+        [JsonProperty("reason")]
+        public IncompleteMessageReason Reason { get; private set; }
+    }
+}
