+++ OpenAI/Packages/com.openai.unity/Runtime/Threads/IncompleteDetails.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c07e2d64a9318b54c9c421c2e3d9fb1e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
