+++ OpenAI/Packages/com.openai.unity/Runtime/Threads/IncompleteMessageReason.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f447af0a4dfeb7438c19340cbe03a8f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
