+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoreFileBatch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9eaa0a0400c304f4ab68295b617f846c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
