+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoreFileResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f44d9d3af58f7f443ab0c5f53705c1ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
