+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoreFileStatus.cs
@@ -1,0 +1,17 @@
+using System.Runtime.Serialization;
+
+namespace OpenAI.VectorStores
+{
+    public enum VectorStoreFileStatus
+    {
+        NotStarted = 0,
+        [EnumMember(Value = "in_progress")]
+        InProgress,
+        [EnumMember(Value = "completed")]
+        Completed,
+        [EnumMember(Value = "cancelled")]
+        Cancelled,
+        [EnumMember(Value = "failed")]
+        Expired
+    }
+}
