+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoreFileStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f46d6b6f8c2c604db0d3666cf3fd060
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
