+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoreStatus.cs
@@ -1,0 +1,15 @@
+using System.Runtime.Serialization;
+
+namespace OpenAI.VectorStores
+{
+    public enum VectorStoreStatus
+    {
+        NotStarted = 0,
+        [EnumMember(Value = "in_progress")]
+        InProgress,
+        [EnumMember(Value = "completed")]
+        Completed,
+        [EnumMember(Value = "failed")]
+        Expired
+    }
+}
