+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoreStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4d2d09c9a8d2115478cd0feab62607ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
