+++ OpenAI/Packages/com.openai.unity/Runtime/VectorStores/VectorStoresEndpoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ca4585db4a194d4ca54038728778e5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 84a7eb8fc6eba7540bf56cea8e12249c, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
